Remove PetabImporterPysb (#1082)
Since https://github.com/ICB-DCM/pyPESTO/pull/1073, we don't need an extra `PetabImporterPysb`, as anything PySB-specific is handled inside PEtab/AMICI.
