+++ pypesto/petab/__init__.py
@@ -7,7 +7,6 @@ pyPESTO support for the PEtab data format.
 import warnings
 
 from .importer import PetabImporter
-from .pysb_importer import PetabImporterPysb
 
 # PEtab and amici are optional dependencies
